requirements.txt
# Core
django==4.2.8
python-dotenv==1.0.0
str2bool==1.1

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0

django-allauth==0.58.1
pillow==10.3.0
django-webpack-loader==3.1.0
django-autoslug==1.9.9
django-quill-editor==0.1.40
djangorestframework==3.15.1
django-filter==24.2
django-cors-headers==4.3.1

celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1

# Tools (generator, deployer, GH)
black
gitpython
wget
requests
PyGithub==1.55

# DOCS
Markdown==3.6
drf-yasg==1.21.7

# DOCS Sphinx = https://www.515tech.com/post/integrating-sphinx-documentation-into-your-django-project
sphinx
django-docs
sphinx-rtd-theme
recommonmark

psycopg2-binary==2.9.9

sphinx-book-theme
sphinx-copybutton
sphinxcontrib-googleanalytics==0.4
markdown2
html2text